added confirm dialog boxes

+++ src/components/book/SingleBook.jsx
@@ -17,8 +17,14 @@ export default function SingleBook({ book }) {
 
     const deleteBook = async (e, id) => {
         e.preventDefault();
+        const confirmDelete = window.confirm('Are you sure you want to delete this book?');
+
+        if (!confirmDelete) return;
         await deleteDocument('books', id);
-        await deleteFileFromStorage(`/covers/${user.uid}/${book.bookCoverName}`);
+        if (book.bookCoverName) {
+            await deleteFileFromStorage(`/covers/${user.uid}/${book.bookCoverName}`);
+            console.log('good')
+        }
     };
 
     const toggleSaved = async (e, bookId) => {
